requirements.txt
Django==2.1.1
djangorestframework==3.10.2
pytz==2019.2
mysqlclient==1.3.12
